Local server
Heroku needs the "npm install -g serve" because that package turns the working directory into a virtual directory, so it is hosted in localhost. Now npm start is linked to "serve -s build" because that way the build is essentially hosted by a local prod server, instead of the dev server of npm start, which why now npm run devstart is linked to npm start. Another way of doing this would have been to write a small express server